+++ src/pages/Editprof.jsx
@@ -4,15 +4,57 @@ import { useCookies } from "react-cookie";
 import "./editprof.scss";
 import { Header } from "../components/Header";
 import { url } from "../env";
+// import resizeImage from "../components/ResizeImage";
 
 export function Editprof() {
   const [errorMessage, setErrorMessage] = useState("");
-  const [successMessage, setSuccessMessage] = useState("");
+  const [successMessage,setSuccessMessage ] = useState("");
   const [name, setName] = useState("");
   const [currentIcon, setCurrentIcon] = useState("");
-  const [/*editIcon*/, setEditIcon] = useState("");
   const [cookies] = useCookies();
   const [previewImage, setPreviewImage] = useState(null);
+  const [, setEditIcon] = useState("");
+  // const [, setCookie] = useCookies();
+
+  const handleSubmit = async (event) => {
+    event.preventDefault();
+    try {
+        await axios.put(
+      // const response = await axios.put(
+        `${url}/users`,
+        { name },
+        {
+          headers: {
+            authorization: `Bearer ${cookies.token}`,
+          },
+        },
+      );
+      // const token = response.data.token;
+      // setCookie("token", token);
+
+      // const resizedImage = await resizeImage(editIcon, 300, 300);
+      // const formData = new FormData();
+      // formData.append("icon", resizedImage, resizedImage.name);
+
+      // const iconResponse = await axios.post(`${url}/uploads`, formData, {
+      //   headers: {
+      //     Authorization: `Bearer ${token}`,
+      //     "Content-Type": "multipart/form-data",
+      //   },
+      // });
+      // if (iconResponse.status === 200) {
+      //   // navigate("/");
+      //   console.log("画像登録完了", iconResponse);
+      // } else {
+      //   console.error("Upload failed");
+      //   // setError("アップロードに失敗しました。");
+      // }
+      // setPostComplete(true);
+      setSuccessMessage(`ユーザー名の更新ができました`);
+    } catch (err) {
+        console.error("APIリクエストエラー", err);
+      }
+    };
 
   const fetchUserInfo = async () => {
     try {
@@ -25,7 +67,6 @@ export function Editprof() {
       console.log("ユーザー情報を表示します", user_info);
       setName(user_info.name);
       setCurrentIcon(user_info.iconUrl);
-
     } catch (err) {
       setErrorMessage(`ユーザー情報の取得に失敗しました。${err}`);
     }
@@ -35,27 +76,6 @@ export function Editprof() {
   useEffect(() => {
     fetchUserInfo();
   }, [cookies.token]);
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-
-    try {
-      await axios.put(
-        `${url}/users`,
-        { name },
-        {
-          headers: {
-            authorization: `Bearer ${cookies.token}`,
-          },
-        },
-      );
-      // PUT リクエストの結果を取得
-      console.log(`更新が完了しました`);
-      setSuccessMessage(`ユーザー名を"${name}"に更新できました。\nアイコンを"${previewImage}"に更新できました`);
-    } catch (error) {
-      setErrorMessage(`更新に失敗しました。${error}`);
-    }
-  };
 
   const handlePhotoChange = (event) => {
     const selectedPhoto = event.target.files[0];
@@ -78,7 +98,7 @@ export function Editprof() {
         <form id="edit-profile">
           <label htmlFor="name">ユーザー名:</label>
           <input
-            className="name-input"
+            className="name-input1"
             type="text"
             id="name"
             name="name"
@@ -86,25 +106,26 @@ export function Editprof() {
             onChange={(e) => setName(e.target.value)}
           />
           <div>
-          <label htmlFor="photo">登録済みアイコン:</label>
-          <div>
+            <label htmlFor="photo">登録済みアイコン:</label>
+            <div>
               <img
                 src={currentIcon}
                 alt="登録済みアイコン"
                 className="uploaded-image"
               />
-          </div>
-          <label htmlFor="photo">変更後アイコン:<input
-              type="file"
-              id="photo"
-              name="photo"
-              accept="image/*"
-              onChange={handlePhotoChange}
-              required
-            /></label>
-          <div>
-
             </div>
+            <label htmlFor="photo">
+              変更後アイコン:
+              <input
+                type="file"
+                id="photo"
+                name="photo"
+                accept="image/*"
+                onChange={handlePhotoChange}
+                required
+              />
+            </label>
+            <div></div>
             {previewImage && (
               <img
                 src={previewImage}
