+++ src/pages/SignUp.jsx
@@ -5,7 +5,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { signIn } from "../authSlice";
 import "./signup.scss";
-import Compressor from "image-compressor.js";
+import resizeImage from "../components/ResizeImage";
 import { url } from "../env";
 
 export function SignUp() {
@@ -25,22 +25,7 @@ export function SignUp() {
       navigate("/");
     }
   });
-  const resizeImage = (file, maxWidth, maxHeight) => {
-    return new Promise((resolve, reject) => {
-      const compressor = new Compressor();
 
-      compressor.compress(file, {
-        maxWidth,
-        maxHeight,
-        success(result) {
-          resolve(result);
-        },
-        error(err) {
-          reject(err);
-        },
-      });
-    });
-  };
   const handleSubmit = async (event) => {
     event.preventDefault(); // ページがリロードされないようフォームのデフォルトの動作を防止
     if (!name || !email || !password) {
@@ -72,7 +57,7 @@ export function SignUp() {
         console.error("Upload failed");
         setError("アップロードに失敗しました。");
       }
-      // setPostComplete(true);
+
     } catch (err) {
       if (err.response && err.response.status === 409) {
         // ステータスコード409の場合、ユーザーがすでに登録されているとみなす
