+++ src/pages/editprof.scss
@@ -1,6 +1,6 @@
 @charset "utf-8";
 
-.name-input {
+.name-input1 {
   width: 10%;
   font-size: large;
   height: 30px;
